Move confirm logic out of StateManager
StateManager contains the code to execute upon command confirmation.

While this was done to keep state management tightly coupled
with the execution of the pending operation, this brings much
logic into State.

Extract logic in state into ConfirmCommand class. Instead of
handling state update in StateManager itself, expose a method
to update state and pass the updating of state as a callback
to ConfirmCommand.

This way, ConfirmCommand holds all related-logic and
state updating can still be controlled.