Improve delete tests
Tests for delete class simply check if the confirm
command is returned upon creation of a delete command.

Change some test suite names to reflect that no
deletion is done and only a confirmation prompt is made.

Add methods to TestUtil to allow for testing results
after the confimation prompt is handled.

Add tests to delete to check that persons are deleted
after confirming.