Refactor CommandHistory, Add saved history to file
Move CommandHistory to model component and add methods such that
ModelManager mostly handles CommandHistory.

Add Storage for CommandHistory component and update
Storage Manager and interfaces.

Move CommandHistory management logic to Logic component. Add
storage management of CommandHistory in Logic.

Update UIManager constructor to take in CommandHistory from Model.
Add methods to inject CommandHistory into the UI component. This
was done instead of introducing the model component to the UI so
as to maintain the loose level of coupling that the UI component
has.

Overall, these changes were done to better follow the current
structure and architecture of the software.