Refactor CommandPendingResult supplier parameter
ConfirmationPendingResult accepts a supplier for CommandResult.
However, developers can cause an issue if the supplier is haphazardly
called multiple times, or if the return result is invalid (null).

Separate the supplier into two components, a runnable that should
be ran on confirmation and a command result that should return on
success. Wrap the runnable in a OnceRunnable class to ensure
that the callback can only be ran once. Add assertion to prevent
misuse. Update tests.