Improve code quality of List Command and Parser
Improved readbility of conditional statements in ListCommandParser by
utilising the created boolean flags instead of re-calling the the
isPresent() method again.

Refactored the execute method in ListCommand to improve SLAP by
extracting the logic for getting the result message into the SortOrder
enum.

Repositioned SortOrder enum to the bottom of the file.

Decided to keep SortOrder enum as a nested public enum inside
ListCommand. This is because it is highly cohesive with ListCommand as
it is only used by this command. Hence, keeping them together maximizes
the command's internal coherence. Furthermore, this reduces coupling.