Fixed LaunchCommand by using DesktopApi solely
This fixes the issue for some specific linux distro crashing even though
they support the java.awt.Desktop api, it still seems to crash. Hence,
by using the DesktopApi class created instead, it will first try to
launch based on the system specific os. And if that fails it will
attempt to launch via java.awt.Desktop api (ONLY IF IT IS SUPPORTED).
Hence as mentioned earlier, there is still a possibility that if a linux
distro is able reach that point, that the code would crash.