Change Help Window
Instead of lauching a smaller modal, users will be immediately
redirected to the user guide web page. This removes the current known
issue that occurs when minimizing the help window.

As a result, I have since removed all references to HelpWindow and have
deleted HelpWindow ui and resource files. Instead handleHelp() method in
MainWindow will immediately trigger a launchUserGuide method in
ApplicationLinkLauncher which in turn launches the web page of the user
guide.